writer: refactor
Break up the implementation of the writer and translator. This is to
prepare for prospect support for alternative translators.

     +--------+    +------------+
     | Writer |----+ Translator |
     +--------+    +------^-----+
                          |
                 +----------------+
                 | WikiTranslator |
                 +----------------+

Files:

 translator/shared.py | common translator logic
 translator/wiki.py   | wiki format translator
 writer.py            | confluence builder's sphinx writer
